fix: Use context.z position has an offset from the calculated display z position in Aux Builder (see description).
Making context.z act as an offset allows context.z value of 0 to place the file on the “ground” regardless of tile height in Aux Builder and always place the file on the ground in Aux Builder. No more file clipping issues due to grid planes being at different heights in Aux Builder and Aux Player.