fix: An issue where some bots would not be added to the pagePortal if one of those new bots defines the same portal
Issue was that there were some bots that defined another pagePortal at the same dimension so some bots would see themselves as already being in the dimension and would be skipped.
Fix works by letting the BotDimensionManager skip some bots so they won't be evaluated for defining the portal and will therefore not pre-add the bots.