fix: An issue where creating a bot that immediately starts a timer could encounter issues with reading updated tag values
This stems from a consequence of a design choice where if a new bot is created, the relevant events will be sent to the correct partition and then the new bot will be sent back to the runtime from the partition. However, the runtime already has its own (non synced) version of the bot, so it has to replace that bot with the bot from the partition. If scripts are able to get a reference to this old bot, then tag updates will seem to stop working.