feat: Add XpStore and implementation in MemoryStore
This commit adds the XpStore interface and related interfaces for managing XP accounts, users, contracts, invoices, and account entries. It also includes the implementation of these interfaces in the MemoryStore class. The XpStore allows saving and retrieving XP accounts and users, as well as performing transactions and accessing meta data associated with users in the Xp system.