feat: Add XpController and framework methods
This commit adds the implementation of the createXpAccount and createXpUser methods in the XpController class. The createXpAccount method creates a new XP account with a unique ID and initializes it with default values. The createXpUser method creates a new XP user with a unique ID and associates it with an email address. Both methods are asynchronous and will be updated but are not yet fully able to perform the necessary operations to save the created accounts and users in the XpStore.