refactor: Update docker-compose.dev.yml to enable IPv6 networking
This commit updates the docker-compose.dev.yml file to enable IPv6 networking. It adds a new network called enable_ipv6 with the bridge driver and assigns IPv6 subnets. The minio and redis services are now configured to use this network.

Refactor the docker-compose.dev.yml file to enable IPv6 networking and reorganize the service configurations.